pyproject.toml
[project]
name = "volume-optimizer"
version = "1.0.0"
description = "Professional API for optimizing hypertrophy training volume with tiered subscriptions and advanced analytics."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.19.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
    "httpx>=0.26.0",
    "pytest-asyncio>=0.23.0",
]
